Fix Console breaking focus navigation
# Problem
ConsoleLine breaks focus navigation with `ui_up` and `ui_down`

# Test scenario
Create a simple scene with a `vBoxContainer` and 3 `button` children. Click any button and try changing the focused button with `up` and `down` arrows. Focus works as long as only `up` or `down` is pressed, but the buttons lose focus when `down` is pressed after pressing `up`

# Fix
Proposed fix simply returns early when Console is not visible. Other possibility is to toggle `set_process_input()` based on visibility changes.