Reimplement History without QueueCollection
The QueueCollection stuff is buggy and not necessary, use an array instead.

The misc/collection stuff uses a dictionary for a queue, with consecutive integers stored in the dict keys instead of indexing, I assume its some holdover from early versions of Godot. Once items are removed from the start of the queue rather than the end (e.g. if maxLength exceeded) then the length of the dictionary ends up less than the largest key/"index" and a lot of the collection functions break.